+++ packages/segmented-control/src/index.ts
@@ -1,6 +1,4 @@
-export { SegmentedControl } from './SegmentedControl';
-export { SegmentedControlOption } from './SegmentedControlOption';
-export type {
-  SegmentedControlProps,
-  SegmentedControlOptionProps,
-} from './types';
+export { SegmentedControl } from './SegmentedControl/SegmentedControl';
+export { SegmentedControlOption } from './SegmentedControlOption/SegmentedControlOption';
+export type { SegmentedControlProps } from './SegmentedControl/types';
+export type { SegmentedControlOptionProps } from './SegmentedControlOption/types';
